Subclasses are now tracked
TracePoint returns the original class when a sublclass is called.

For example:

```ruby
class Foo
  def split(string)
  end
end

class Bar < Foo
end

class Zar < Bar
end

TracePoint.trace(:call, :c_call) do |tp|
  tp.disable
  puts "=="
  puts tp.defined_class
  puts tp.method_id
  puts caller
  tp.enable
end

Zar.new.split("")
```

Will return

```
# Foo
# split
# scratch.rb:2:in `split'
# scratch.rb:21:in `<main>'
```

Even though we called `Zar`.

To get around this we can loop through the caller and determine the first location where the method `split` is actually called. This is normally `caller.first` but in cases of subclasses it would be `caller[0]`. 

It is strange that it returns the method definition location in the first caller location, I would think it would be the other way around. Either way this is much more durable and doesn't strictly rely on the the order of the caller.